:lipstick: Update my input UI
